serverspec result tabpane
